# backend/internal/pkg/api/timeline_service_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v7.34.1
// source: timeline_service.proto

package api

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationTimelineServiceGetPhotosByTimeline = "/api.TimelineService/GetPhotosByTimeline"
const OperationTimelineServiceListTimelines = "/api.TimelineService/ListTimelines"

type TimelineServiceHTTPServer interface {
	// GetPhotosByTimeline 某时间线下的照片
	GetPhotosByTimeline(context.Context, *GetPhotosByTimelineRequest) (*GetPhotosByTimelineResponse, error)
	// ListTimelines 所有时间线列表
	ListTimelines(context.Context, *Empty) (*ListTimelinesResponse, error)
}

func RegisterTimelineServiceHTTPServer(s *http.Server, srv TimelineServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/v1/timelines", _TimelineService_ListTimelines0_HTTP_Handler(srv))
	r.GET("/api/v1/timelines/{name}/photos", _TimelineService_GetPhotosByTimeline0_HTTP_Handler(srv))
}

func _TimelineService_ListTimelines0_HTTP_Handler(srv TimelineServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTimelineServiceListTimelines)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListTimelines(ctx, req.(*Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListTimelinesResponse)
		return ctx.Result(200, reply)
	}
}

func _TimelineService_GetPhotosByTimeline0_HTTP_Handler(srv TimelineServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetPhotosByTimelineRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationTimelineServiceGetPhotosByTimeline)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetPhotosByTimeline(ctx, req.(*GetPhotosByTimelineRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetPhotosByTimelineResponse)
		return ctx.Result(200, reply)
	}
}

type TimelineServiceHTTPClient interface {
	// GetPhotosByTimeline 某时间线下的照片
	GetPhotosByTimeline(ctx context.Context, req *GetPhotosByTimelineRequest, opts ...http.CallOption) (rsp *GetPhotosByTimelineResponse, err error)
	// ListTimelines 所有时间线列表
	ListTimelines(ctx context.Context, req *Empty, opts ...http.CallOption) (rsp *ListTimelinesResponse, err error)
}

type TimelineServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewTimelineServiceHTTPClient(client *http.Client) TimelineServiceHTTPClient {
	return &TimelineServiceHTTPClientImpl{client}
}

// GetPhotosByTimeline 某时间线下的照片
func (c *TimelineServiceHTTPClientImpl) GetPhotosByTimeline(ctx context.Context, in *GetPhotosByTimelineRequest, opts ...http.CallOption) (*GetPhotosByTimelineResponse, error) {
	var out GetPhotosByTimelineResponse
	pattern := "/api/v1/timelines/{name}/photos"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTimelineServiceGetPhotosByTimeline))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// ListTimelines 所有时间线列表
func (c *TimelineServiceHTTPClientImpl) ListTimelines(ctx context.Context, in *Empty, opts ...http.CallOption) (*ListTimelinesResponse, error) {
	var out ListTimelinesResponse
	pattern := "/api/v1/timelines"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationTimelineServiceListTimelines))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}