# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/timelines/{name}/photos:
        get:
            tags:
                - TimelineService
            description: 某时间线下的照片
            operationId: TimelineService_GetPhotosByTimeline
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.GetPhotosByTimelineResponse'
...
components:
    schemas:
...
        api.GetPhotosByTimelineResponse:
            type: object
            properties:
                timeline:
                    type: string
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.PhotoItem'
                total:
                    type: integer
                    format: int32
...
        api.PhotoItem:
            type: object
            properties:
                id:
                    type: string
                filename:
                    type: string
                filePath:
                    type: string
                timeline:
                    type: string
                tags:
                    type: string
                description:
                    type: string
                objects:
                    type: string
                colors:
                    type: string
                scene:
                    type: string
                lighting:
                    type: string
                mood:
                    type: string
                composition:
                    type: string
                shotAt:
                    type: string
                width:
                    type: integer
                    format: int32
                height:
                    type: integer
                    format: int32
                brand:
                    type: string
                model:
                    type: string
                lens:
                    type: string
                focalLength:
                    type: string
                aperture:
                    type: string
                iso:
                    type: integer
                    format: int32
                exposureTime:
                    type: string
                latitude:
                    type: number
                    format: double
                longitude:
                    type: number
                    format: double
                altitude:
                    type: number
                    format: double
                importedAt:
                    type: string
                hasDescription:
                    type: boolean
                thumbnailUrl:
                    type: string
...
servers:
    - description: PGO API
      url: http://127.0.0.1:8080